chore(Icon): UXD-756 extract template
