+++ packages/Icon/svgr.config.js
@@ -1,3 +1,5 @@
+const template = require("./helper/template");
+
 module.exports = {
   icon: true,
   replaceAttrValues: {
@@ -15,16 +17,5 @@ module.exports = {
     focusable: false,
   },
   titleProp: true,
-  template: ({ template }, opts, { imports, interfaces, componentName, jsx, exports }) => {
-    const plugins = ["jsx"];
-    const typeScriptTpl = template.smart({ plugins });
-    return typeScriptTpl.ast`
-    ${imports}
-    ${interfaces}
-    function ${componentName}({title=null, ...props}) {
-      return ${jsx};
-    }
-    ${exports}
-  `;
-  },
+  template,
 };
